Move ping thing charts to a separate directory and extract shared variables
