+++ app/javascript/packs/ping_things/charts/chart_variables.js
@@ -1,0 +1,5 @@
+export default {
+    chart_darkgrey: '#979797',
+    chart_grid_color: '#322f3d',
+    chart_font_size: 11,
+}
